web_site/requirements.txt
flask-mongoengine
Flask-Script
